pyproject.toml
[project]
name = "isg-science-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "transformers",
    "anthropic>=0.68.0",
    "anytree>=2.13.0",
    "backoff>=2.2.1",
    "black>=25.9.0",
    "chardet>=5.2.0",
    "dataclasses-json>=0.6.7",
    "dataloader",
    "faiss-cpu>=1.13.2",
    "funcy>=2.0",
    "graphviz>=0.21",
    "humanize>=4.13.0",
    "jsonschema>=4.25.1",
    "langchain>=1.2",
    "langchain-community>=0.4.1",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.3",
    "lenskit==2025.6.2",
    "matplotlib>=3.10.7",
    "mcp[cli]>=1.24.0",
    "numpy>=1.26.4",
    "openai>=1.108.1",
    "omnirec",
    "pydantic-settings>=2.10.1",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.0",
    "rich>=14.1.0",
    "scikit-learn==1.7.1",
    "shutup>=0.2.0",
    "tomli-w>=1.2.0",
]

[tool.uv.workspace]
members = ["packages/dataloader"]

[tool.uv.sources]
dataloader = { workspace = true }
omnirec = { git = "https://github.com/ISG-Siegen/OmniRec.git", branch = "develop" }